requirements.txt
Django==4.1.7
python-telegram-bot
